Break up report_listing.tsx into Default, Stateful, and ListingTable
