+++ x-pack/plugins/reporting/public/management/index.ts
@@ -8,20 +8,21 @@
 import type { ApplicationStart, ToastsSetup } from '@kbn/core/public';
 import type { LicensingPluginStart } from '@kbn/licensing-plugin/public';
 import type { ClientConfigType } from '@kbn/reporting-public';
-import type { UseIlmPolicyStatusReturn } from '../lib/ilm_policy_status_context';
 import type { ReportingAPIClient } from '../lib/reporting_api_client';
 import type { SharePluginSetup } from '../shared_imports';
 
 export interface ListingProps {
   apiClient: ReportingAPIClient;
-  capabilities: ApplicationStart['capabilities'];
   license$: LicensingPluginStart['license$'];
   config: ClientConfigType;
   redirect: ApplicationStart['navigateToApp'];
   navigateToUrl: ApplicationStart['navigateToUrl'];
   toasts: ToastsSetup;
   urlService: SharePluginSetup['url'];
-  ilmPolicyContextValue?: UseIlmPolicyStatusReturn;
 }
+
+export type ListingPropsInternal = ListingProps & {
+  capabilities: ApplicationStart['capabilities'];
+};
 
 export { ReportListing } from './report_listing';
